Move profile options into own card
