+++ src/Basics/Styles.js
@@ -67,6 +67,7 @@ let Styles = {
         color: Colors.textDarkGray
     },
     profileCard: {
+        backgroundColor: "white",
         boxShadow: "0px 4px 16px rgba(0, 0, 0, 0.1)",
         borderRadius: "4px",
         "& > h2": {
